combine cl signals in one file
