+++ lumispy/signals/cl_spectrum.py
@@ -18,8 +18,8 @@
 
 """Signal class for Cathodoluminescence spectral data.
 """
-from inspect import getfullargspec
 
+from inspect import getfullargspec
 import numpy as np
 from warnings import warn
 
@@ -119,6 +119,77 @@ class CLSpectrum(LumiSpectrum):
 
 
 class LazyCLSpectrum(LazySignal, CLSpectrum):
+    _lazy = True
+
+    pass
+
+
+"""SEM specific signal class for Cathodoluminescence spectral data.
+"""
+
+class CLSEMSpectrum(CLSpectrum):
+    _signal_type = "CL_SEM"
+
+    def correct_grating_shift(self, cal_factor_x_axis, corr_factor_grating, sem_magnification, **kwargs):
+        """"
+        Applies shift caused by the grating offset wrt the scanning centre.
+        Authorship: Gunnar Kusch (gk419@cam.ac.uk)
+
+        :param cal_factor_x_axis: The navigation correction factor.
+        :param corr_factor_grating: The grating correction factor.
+        :param sem_magnification: The SEM (real) magnification value.
+            For the Attolight original metadata, take the `SEM.Real_Magnification` value
+        :param kwargs: The parameters passed to `hyperspy.align1D()` function like:
+            interpolation_method ('linear', 'nearest', 'zero', 'slinear', 'quadratic, 'cubic')
+            parallel: Bool
+            crop, expand, fill_value ...
+
+        """
+
+        # Avoid correcting for this shift twice (first time it fails, so except
+        # block runs. Second time, try succeeds, so except block is skipped):
+        try:
+            self.metadata.Signal.grating_corrected == True
+        except AttributeError:
+            # Get all relevant parameters
+            nx = self.axes_manager.navigation_shape[0]
+            ny = self.axes_manager.navigation_shape[1]
+            fov = sem_magnification
+
+            # Correction of the Wavelength Shift along the X-Axis
+            calax = cal_factor_x_axis / (fov * nx)
+            # (Total Variation, Channels, Step)
+            garray = np.arange((-corr_factor_grating / 2) * calax * 1000 * nx,
+                               (corr_factor_grating / 2) * calax * 1000 * nx,
+                                corr_factor_grating * calax * 1000)
+            barray = np.full((ny, nx), garray)
+
+            self.shift1D(barray, **kwargs)
+
+            # Store modification in metadata
+            self.metadata.set_item("Signal.grating_corrected", True)
+        else:
+            raise Exception("You already corrected for the grating shift.")
+
+
+class LazyCLSEMSpectrum(LazySignal, CLSEMSpectrum):
+    _lazy = True
+
+    pass
+
+
+"""STEM specific signal class for Cathodoluminescence spectral data.
+"""
+
+class CLSTEMSpectrum(CLSpectrum):
+
+    _signal_type = "CL_STEM"
+
+    pass
+
+
+class LazyCLSTEMSpectrum(LazySignal, CLSTEMSpectrum):
+
     _lazy = True
 
     pass
